Mask each UART IRQ individually to improve bit-timing accuracy
Instead of globally disabling the lock to achieve mutual exclusion to
the ring buffers, mask each interrupt individually when needed.
This will reduce latency for the IRQ handlers which is critical for the
SWUART implementation.

This also makes the ports "truly" full duplex as each port and each
direction can work independently without being blocked when another
port is queing or popping data.