Offset SWUART RX sampling for more robust sampling
When the startbit is detected, start sampling after half the bittime.
This way sampling will occur in the middle of the UART data cycle and
makes the receiver less vulnerable to inaccurate transmitters.
We can also detect glitches on the RX pin by re-sampling the startbit
after half the bittime, and abort sampling if it has vanished. This
makes the receiver much more robust when running at higher baudrates
and with a higher cpu clockcycle (where no significant delay is added
by the interrupt handler. Baudrate 4800 is now possible!