Heap-based buffers for more finegrain ringbuf sizes
Switch to heap-based memory allocation for the ringbuffer to allow for
individual sizes for each ringbuffer instance. This will save memory
since not all UART buffers need to be as large as the GSM buffer where
we expect more data at a higher rate.

This also means that each ringbuffer instance must be free'd after
usage to prevent memory leaks.